+++ packages/office-ui-fabric-react/src/components/Announced/examples/Announced.LazyLoading.Example.tsx
@@ -11,7 +11,7 @@ import { mergeStyles } from 'office-ui-fabric-react/lib/Styling';
 import { useBoolean, useSetInterval, useConst } from '@uifabric/react-hooks';
 
 const DELAY = 10;
-const PHOTO_COUNT = 30;
+const PHOTO_COUNT = 15;
 
 const stackTokens: IStackTokens = { childrenGap: 10 };
 const photoStackTokens: IStackTokens = { childrenGap: '6 6' };
@@ -19,11 +19,6 @@ const photoStackTokens: IStackTokens = { childrenGap: '6 6' };
 const defaultButtonStyles: Partial<IButtonStyles> = { root: { width: 150 } };
 
 const photoStackStyles: Partial<IStackStyles> = {
-  root: {
-    border: '1px solid black',
-    padding: 10,
-    overflowY: 'auto',
-  },
   inner: {
     padding: 0,
   },
@@ -44,8 +39,7 @@ interface IPhoto {
 
 export const AnnouncedLazyLoadingExample = () => {
   const [total, setTotal] = React.useState<number>(0);
-  const [announced, setAnnounced] = React.useState<JSX.Element | undefined>(undefined);
-  const [timeSinceLastAnnounce, setTimeSinceLastAnnounce] = React.useState<number>(0);
+  const [, setTimeSinceLastAnnounce] = React.useState<number>(0);
   const [loading, { toggle: toggleLoading }] = useBoolean(false);
   const percentComplete = total / PHOTO_COUNT;
   const { setInterval, clearInterval } = useSetInterval();
@@ -53,82 +47,81 @@ export const AnnouncedLazyLoadingExample = () => {
   const photos: IPhoto[] = useConst(() => {
     const width = 100;
     const height = 100;
-    const result = createArray(PHOTO_COUNT, () => {
-      return {
-        url: `http://placehold.it/${width}x${height}`,
-        width: width,
-        height: height,
-      };
-    });
-    return result;
+    return createArray(PHOTO_COUNT, () => ({
+      url: `http://placehold.it/${width}x${height}`,
+      width,
+      height,
+    }));
   });
 
   React.useEffect(() => {
     if (loading) {
       setTotal(0);
+
       const itemIntervalId = setInterval(() => {
         setTotal((t: number) => {
           if (t < PHOTO_COUNT) {
             return t + 1;
-          } else {
-            clearInterval(itemIntervalId);
-            return t;
           }
+          clearInterval(itemIntervalId);
+          toggleLoading();
+          return t;
         });
       }, 500);
 
       const announceIntervalId = setInterval(() => {
-        setTimeSinceLastAnnounce(timeSinceLastAnnounce + 1);
-        if (timeSinceLastAnnounce === DELAY || total === PHOTO_COUNT) {
-          setAnnounced(<Announced message={`${total} of ${PHOTO_COUNT} photos loaded`} />);
-          setTimeSinceLastAnnounce(0);
-          if (total === PHOTO_COUNT) {
-            clearInterval(announceIntervalId);
+        setTimeSinceLastAnnounce(timeSinceLastAnnounce => {
+          if (timeSinceLastAnnounce === DELAY || total === PHOTO_COUNT) {
+            if (total === PHOTO_COUNT) {
+              clearInterval(announceIntervalId);
+            }
+            return 0;
           }
-        }
-      }, 1000);
+          return timeSinceLastAnnounce + 1;
+        });
+      }, 2000);
 
       return () => {
         clearInterval(itemIntervalId);
         clearInterval(announceIntervalId);
       };
     }
-  }, [clearInterval, setInterval, timeSinceLastAnnounce, loading, total]);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [clearInterval, setInterval, loading]);
 
   return (
-    <Stack tokens={stackTokens}>
-      <Text>
-        Turn on Narrator and press the button to start loading photos. Announced should announce the number of photos
-        loaded every 10 seconds, as that is the delay chosen for this example.
-      </Text>
-      <DefaultButton text={loading ? 'Cancel' : 'Load photos'} onClick={toggleLoading} styles={defaultButtonStyles} />
-      <ProgressIndicator
-        label={percentComplete < 1 ? 'Loading photos' : 'Finished loading photos'}
-        percentComplete={percentComplete}
-      />
-      {announced}
-      <FocusZone>
-        <Stack
-          horizontal
-          wrap
-          tokens={photoStackTokens}
-          styles={photoStackStyles}
-          slots={{ inner: { component: 'ul' } }}
-        >
-          {photos.map((photo: IPhoto, index: number) => (
-            <li
-              key={index}
-              className={photoCellClass}
-              aria-posinset={index + 1}
-              aria-setsize={PHOTO_COUNT}
-              aria-label="Photo"
-              data-is-focusable
-            >
-              {total > index ? <Image src={photo.url} width={photo.width} height={photo.height} /> : <div />}
-            </li>
-          ))}
-        </Stack>
-      </FocusZone>
-    </Stack>
+    <>
+      <Stack tokens={stackTokens}>
+        <Text>
+          Turn on Narrator and press the button to start loading photos. The number of photos loaded will be announced
+          every 2 seconds.
+        </Text>
+        <DefaultButton text={loading ? 'Cancel' : 'Load photos'} onClick={toggleLoading} styles={defaultButtonStyles} />
+        <ProgressIndicator label={!loading ? 'Paused' : 'Loading photos'} percentComplete={percentComplete} />
+        <FocusZone>
+          <Stack
+            horizontal
+            wrap
+            tokens={photoStackTokens}
+            styles={photoStackStyles}
+            slots={{ inner: { component: 'ul' } }}
+          >
+            {photos.slice(0, total).map((photo: IPhoto, index: number) => (
+              <li
+                key={index}
+                className={photoCellClass}
+                aria-posinset={index + 1}
+                aria-setsize={PHOTO_COUNT}
+                aria-label="Photo"
+                data-is-focusable
+              >
+                <Image src={photo.url} width={photo.width} height={photo.height} />
+              </li>
+            ))}
+          </Stack>
+        </FocusZone>
+      </Stack>
+      {loading && <Announced message={`${total} of ${PHOTO_COUNT} photos loaded`} />}
+    </>
   );
 };