+++ packages/office-ui-fabric-react/src/components/Toggle/Toggle.styles.ts
@@ -1,4 +1,9 @@
-import { HighContrastSelector, getFocusStyle, FontWeights } from '../../Styling';
+import {
+  HighContrastSelector,
+  getFocusStyle,
+  FontWeights,
+  getEdgeChromiumNoHighContrastAdjustSelector,
+} from '../../Styling';
 import { IToggleStyleProps, IToggleStyles } from './Toggle.types';
 
 const DEFAULT_PILL_WIDTH = 40;
@@ -129,8 +134,9 @@ export const getStyles = (props: IToggleStyleProps): IToggleStyles => {
                 },
               ],
               [HighContrastSelector]: {
-                backgroundColor: 'WindowText',
+                backgroundColor: 'Highlight',
               },
+              ...getEdgeChromiumNoHighContrastAdjustSelector(),
             },
           },
         ],